changed from selectAll.users to users.selectAll better readablity and syntax?
